refactor: 이미지 자르기 훅으로 로직 분리
